requirements.txt
attrs==23.2.0 ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
bson==0.5.10 ; python_version >= "3.8" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "4.0"
deepdiff==6.7.1 ; python_version >= "3.8" and python_version < "4.0"
dicom-validator==0.3.5 ; python_version >= "3.8" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.8" and python_version < "4.0"
dotty-dict==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.8" and python_version < "4.0"
fhir-resources==7.0.2 ; python_version >= "3.8" and python_version < "4.0"
flywheel-gear-toolkit==0.6.18 ; python_version >= "3.8" and python_version < "4.0"
flywheel-gears==0.3.1 ; python_version >= "3.8" and python_version < "4.0"
flywheel-sdk==17.8.0 ; python_version >= "3.8" and python_version < "4.0"
fw-file==2.4.1 ; python_version >= "3.8" and python_version < "4.0"
fw-meta==3.3.2 ; python_version >= "3.8" and python_version < "4.0"
fw-utils==4.4.1 ; python_version >= "3.8" and python_version < "4.0"
idna==3.7 ; python_version >= "3.8" and python_version < "4.0"
importlib-resources==6.4.0 ; python_version >= "3.8" and python_version < "3.9"
jsonschema-specifications==2023.12.1 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.8" and python_version < "4.0"
memoization==0.4.0 ; python_version >= "3.8" and python_version < "4"
natsort==8.4.0 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.9"
numpy==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
ordered-set==4.1.0 ; python_version >= "3.8" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.8" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.8" and python_version < "4.0"
pathvalidate==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
pydantic==1.10.17 ; python_version >= "3.8" and python_version < "4.0"
pydantic[email]==1.10.17 ; python_version >= "3.8" and python_version < "4.0"
pydicom==2.4.4 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.8" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.8" and python_version < "4.0"
rfc3987==1.3.8 ; python_version >= "3.8" and python_version < "4.0"
rpds-py==0.19.1 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.8" and python_version < "4.0"
tzlocal==5.2 ; python_version >= "3.8" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.19.2 ; python_version >= "3.8" and python_version < "3.9"
